fix(auth): sign out a browser that signed in by cookie
`Auth.logout()` resolved the token through `getBearerToken()`, which read
only the Authorization header. A browser signed in by cookie — which is
what `SocialCallbackAction` produces, and the entire reason
`cookie-auth.ts` exists — carries no such header, so `POST /logout`
revoked nothing, cleared no cookie, and answered 200. Social sign-in could
not be signed out at all: the token stayed valid for its full lifetime
while the API reported success.

The same header-only extraction was hand-rolled a second time in
`authMiddleware`, so the same browser got 401 "No authentication token
provided" on every protected route. Two copies of one lookup is why they
drifted, so both now call `requestToken()`, which checks the header first
and falls back to the auth cookie. Nothing that works today changes; a
credential that was previously invisible now resolves.

That also makes `Auth.user()` and `Auth.currentAccessToken()` work on a
server-rendered request, which were failing the same way.

Second defect found on the way: `Auth.revokeToken()` revoked the access
row only. The refresh cascade lived inline in `logout()` and therefore ran
only for the bearer path, so every other caller — `logoutCookie()` above
all — left a live refresh token behind, and a leaked-but-not-yet-rotated
refresh could mint a new access token after sign-out. That is precisely
the hole `logout()`'s docstring says it exists to close, so the cascade
moves into `revokeToken()` where all callers get it, and into
`revokeAllTokens()` to match the standalone `revokeAllTokens` in tokens.ts
which already cascaded.

`LogoutAction` and `LogoutAllAction` now clear the cookie. Revoking and
clearing are both needed: revoking alone leaves the browser presenting a
dead cookie on every request, clearing alone leaves a copied cookie valid
for the token's lifetime.

`cookie-auth.ts` is split so `authentication.ts` can read a cookie without
importing it back: the value/naming/parsing half moves to `cookie.ts`,
which touches no `Auth`, and `cookie-auth.ts` re-exports it so the public
surface is unchanged.

Verified: auth suite 300 pass / 0 fail, root suite 318 pass / 0 fail,
typecheck 13 (baseline), pickier clean. New `request-token.test.ts` pins
the resolution order, including that the header still wins when a request
carries both.

Refs #2306
