+++ storage/framework/core/auth/src/cookie.ts
@@ -1,0 +1,226 @@
+import { config } from '@stacksjs/config'
+import { log } from '@stacksjs/logging'
+
+/**
+ * Auth cookie values: naming, attributes, and parsing.
+ *
+ * Split out of `cookie-auth.ts` so `authentication.ts` can read the cookie
+ * without importing it back. Nothing here touches `Auth`, which is what makes
+ * that safe; the parts that resolve a token to a user stay in `cookie-auth.ts`.
+ *
+ * The prose explaining WHY the cookie exists, and why it is httpOnly +
+ * SameSite=Lax + conditionally Secure, lives on the re-export in
+ * `cookie-auth.ts`.
+ */
+
+export interface AuthCookieOptions {
+  /** Cookie name. Defaults to {@link authCookieName} — `config.auth.cookie.name`, else `auth-token`. */
+  name?: string
+  /** Lifetime in seconds. Defaults to the configured token expiry. */
+  maxAge?: number
+  /** Path the cookie is sent for. Defaults to `/`. */
+  path?: string
+  /** Domain, for sharing one session across subdomains. */
+  domain?: string
+  /**
+   * Send only over HTTPS. By default this is decided from the configured app
+   * URL: Secure everywhere except a plain-HTTP loopback host (localhost,
+   * `*.localhost`, 127.0.0.1), so a deployment cannot lose the flag by
+   * calling its environment `development`.
+   */
+  secure?: boolean
+  sameSite?: 'Strict' | 'Lax' | 'None'
+}
+
+/**
+ * RFC 6265 cookie-name grammar: a `token`, i.e. any US-ASCII character except
+ * CTLs and separators. A space, comma, semicolon or equals sign makes the
+ * `Set-Cookie` header malformed.
+ */
+const COOKIE_NAME_RE = /^[!#$%&'*+\-.^_`|~0-9A-Za-z]+$/
+
+/**
+ * The one name the auth cookie has.
+ *
+ * There used to be two. `authCookie()` wrote `stacks_auth` (via
+ * `config.auth.cookie.name`, a key that did not exist on `AuthOptions`, so no
+ * app using `satisfies AuthConfig` could even set it), while the Auth
+ * middleware, `team.ts` and the stx page gate all read
+ * `config.auth.defaultTokenName` — `auth-token`. A cookie the framework wrote
+ * was never one the framework read, which is why apps ended up hand-writing a
+ * token pack into `localStorage` from an inline script instead
+ * (stacksjs/stacks#2236).
+ *
+ * Resolution order:
+ *   1. an explicit `options.name`
+ *   2. `config.auth.cookie.name` — the supported key
+ *   3. `config.auth.defaultTokenName` — DEPRECATED, honoured so an app that
+ *      had renamed it (and thereby renamed the cookie those readers wanted)
+ *      keeps working. Ignored with a warning when it is not a legal cookie
+ *      name, which it very often is not: it is a human-readable token label
+ *      like `Web Session`.
+ *   4. `auth-token`
+ */
+export function authCookieName(options?: AuthCookieOptions): string {
+  if (options?.name)
+    return options.name
+
+  const configured = (config.auth as any)?.cookie?.name
+  if (typeof configured === 'string' && configured.length > 0)
+    return configured
+
+  const legacy = (config.auth as any)?.defaultTokenName
+  if (typeof legacy === 'string' && legacy.length > 0 && legacy !== 'auth-token') {
+    if (COOKIE_NAME_RE.test(legacy))
+      return legacy
+
+    // Falling back rather than emitting a malformed Set-Cookie. This is the
+    // failure the overload made likely, so it says what to do about it.
+    console.warn(
+      `[auth] config.auth.defaultTokenName ("${legacy}") is not a valid cookie name and is being ignored `
+      + `for cookie naming; using "auth-token". defaultTokenName is a personal access token label, not a `
+      + `cookie name — set config.auth.cookie.name instead (stacksjs/stacks#2236).`,
+    )
+  }
+
+  return 'auth-token'
+}
+
+function cookieName(options?: AuthCookieOptions): string {
+  return authCookieName(options)
+}
+
+function defaultMaxAge(): number {
+  // `config.auth.tokenExpiry` is milliseconds — the same value
+  // `createTokenForUser` uses for the token's own `expires_at`, so the cookie
+  // and the token it carries die together. A cookie wants seconds.
+  const milliseconds = Number((config.auth as any)?.tokenExpiry ?? 60 * 60 * 1000)
+  if (!Number.isFinite(milliseconds) || milliseconds <= 0)
+    return 60 * 60
+
+  return Math.max(60, Math.round(milliseconds / 1000))
+}
+
+function isLoopbackHost(hostname: string): boolean {
+  const host = hostname.toLowerCase()
+  return host === 'localhost'
+    || host.endsWith('.localhost')
+    || host === '127.0.0.1'
+    || host === '0.0.0.0'
+    || host === '::1'
+    || host === '[::1]'
+}
+
+/**
+ * Whether the auth cookie should carry `Secure`, decided from what the app
+ * demonstrably is rather than what its environment is called.
+ *
+ * The old rule was "Secure unless APP_ENV looks development-ish", and it
+ * failed open: `.env.example` ships `APP_ENV=development`, so an HTTPS
+ * deployment that never changed the env name served its session token
+ * without `Secure` (stacksjs/stacks#2275). Now the URL decides:
+ *
+ * - an `https://` app URL is always Secure
+ * - a plain-HTTP or scheme-less URL drops Secure only on a loopback host
+ *   (localhost, `*.localhost`, 127.0.0.1) — the one place plain HTTP is a
+ *   development reality rather than a misconfiguration
+ * - with no URL configured at all, only the unambiguous `local` / `dev`
+ *   environment names opt out; `development` no longer does
+ *
+ * Exported for tests; `authCookie()` feeds it the live config.
+ */
+export function shouldSecureAuthCookie(
+  app: { url?: unknown, env?: unknown } = (config.app as any) ?? {},
+): boolean {
+  const rawUrl = String(app?.url ?? process.env.APP_URL ?? '').trim()
+  if (rawUrl) {
+    try {
+      const url = new URL(rawUrl.includes('://') ? rawUrl : `http://${rawUrl}`)
+      if (url.protocol === 'https:')
+        return true
+      return !isLoopbackHost(url.hostname)
+    }
+    catch {
+      // An unparseable URL says nothing either way; fall through to the env.
+    }
+  }
+
+  const environment = String(app?.env ?? process.env.APP_ENV ?? '')
+  return !(environment === 'local' || environment === 'dev')
+}
+
+/** Warned once per process, not per response. */
+let warnedInsecureOverride = false
+
+/**
+ * The `Set-Cookie` value that signs a browser in.
+ *
+ * Pair it with the token from `Auth.login()`:
+ *
+ * ```ts
+ * const result = await Auth.login({ email, password })
+ * return new Response(null, {
+ *   status: 303,
+ *   headers: { 'Location': '/account', 'Set-Cookie': authCookie(result.token) },
+ * })
+ * ```
+ */
+export function authCookie(token: string, options: AuthCookieOptions = {}): string {
+  const parts = [
+    `${cookieName(options)}=${encodeURIComponent(token)}`,
+    `Path=${options.path ?? '/'}`,
+    `Max-Age=${options.maxAge ?? defaultMaxAge()}`,
+    'HttpOnly',
+    `SameSite=${options.sameSite ?? 'Lax'}`,
+  ]
+
+  if (options.domain)
+    parts.push(`Domain=${options.domain}`)
+
+  if (options.secure ?? shouldSecureAuthCookie()) {
+    parts.push('Secure')
+  }
+  else if (options.secure === false && shouldSecureAuthCookie() && !warnedInsecureOverride) {
+    // Explicitly stripping Secure off an app whose URL says HTTPS means the
+    // session token is exposed to any plain-HTTP request to the same host.
+    // The default can no longer produce this combination, so it is always a
+    // deliberate override — say so once rather than failing silently open.
+    warnedInsecureOverride = true
+    log.warn('[auth] authCookie() was asked for secure: false while the app URL is HTTPS — the session cookie will also travel over plain HTTP.')
+  }
+
+  return parts.join('; ')
+}
+
+/**
+ * The `Set-Cookie` value that signs a browser out.
+ *
+ * Every attribute except the value has to match the cookie being replaced, or
+ * the browser keeps the original alongside the expired one.
+ */
+export function clearAuthCookie(options: AuthCookieOptions = {}): string {
+  return authCookie('', { ...options, maxAge: 0 })
+}
+
+/** The raw token in a request's auth cookie, if it carries one. */
+export function authCookieToken(request: Request | { headers: Headers }, options: AuthCookieOptions = {}): string | undefined {
+  const header = request.headers.get('cookie')
+  if (!header)
+    return undefined
+
+  const wanted = cookieName(options)
+
+  for (const pair of header.split(';')) {
+    const index = pair.indexOf('=')
+    if (index === -1)
+      continue
+
+    if (pair.slice(0, index).trim() !== wanted)
+      continue
+
+    const value = decodeURIComponent(pair.slice(index + 1).trim())
+    return value.length > 0 ? value : undefined
+  }
+
+  return undefined
+}
