+++ storage/framework/core/auth/src/index.ts
@@ -36,6 +36,7 @@ export * from './session-auth'
 
 // Cookie-carried access tokens, for server-rendered pages.
 export * from './cookie-auth'
+export * from './request-token'
 
 // The stx page gate (`middleware: ['auth' | 'guest']`), token-validating.
 export * from './page-gate'
