+++ storage/framework/core/auth/src/middleware.ts
@@ -1,20 +1,19 @@
 import { Auth } from './authentication'
+import { requestToken } from './request-token'
 
 /**
  * Built-in auth middleware handler
- * Validates bearer token and sets the authenticated user on Auth
+ * Validates the request's access token and sets the authenticated user on Auth
+ *
+ * Resolution is shared with `Auth.getBearerToken()` via `requestToken`, which
+ * checks the Authorization header and then the auth cookie. This used to be a
+ * second hand-rolled copy that stopped at the header, so a browser signed in
+ * by cookie — the whole point of `cookie-auth.ts`, and what
+ * `SocialCallbackAction` produces — was rejected here with 401 on every
+ * protected route (#2306).
  */
 export async function authMiddleware(request: any): Promise<void> {
-  // Try to get bearer token from request method
-  let bearerToken = request.bearerToken?.()
-
-  // Fallback: get directly from Authorization header
-  if (!bearerToken) {
-    const authHeader = request.headers?.get?.('authorization') || request.headers?.get?.('Authorization')
-    if (authHeader && authHeader.startsWith('Bearer ')) {
-      bearerToken = authHeader.substring(7)
-    }
-  }
+  const bearerToken = requestToken(request)
 
   if (!bearerToken) {
     const error = new Error('No authentication token provided.') as Error & { statusCode: number }
